Strange bug creeped into the elasticity material. Corrected it.

+++ Material/pznlmat1d.h
@@ -98,10 +98,10 @@ public:
 	
 protected:
 	/** @brief Cross Section Area */
-	double fArea;
+	REAL fArea;
 	
 	/** @brief Young's modulus */
-	double fE;
+	REAL fE;
 };
 
 #endif
